+++ app/serializers/api/v1/offer_serializer.rb
@@ -9,7 +9,13 @@ class Api::V1::OfferSerializer < ActiveModel::Serializer
   # TODO: https://github.com/cyfronet-fid/marketplace/issues/1964
   attribute :oms_params, if: -> { object.internal? && object.oms_params.present? }
 
+  attribute :bundled_offers, if: -> { object.bundled_offers.present? }
+
   def primary_oms_id
     object.current_oms.id
+  end
+
+  def bundled_offers
+    object.bundled_offers.map(&:slug_iid)
   end
 end