+++ spec/features/order_spec.rb
@@ -217,19 +217,11 @@ RSpec.feature "Service ordering" do
     scenario "I can order bundle resource and its bundled resource" do
       service = create(:service)
       service2 = create(:service)
-      bundled =
-        create(
-          :offer,
-          service: service2,
-          internal: false,
-          order_type: service2.order_type,
-          order_url: service2.order_url
-        )
+      bundled = create(:offer, service: service2, order_type: service2.order_type, order_url: service2.order_url)
       _bundle =
         create(
           :offer,
           service: service,
-          internal: true,
           order_type: service.order_type,
           order_url: service.order_url,
           bundled_offers: [bundled]
@@ -253,7 +245,7 @@ RSpec.feature "Service ordering" do
 
       select "Services", from: "customizable_project_item_project_id"
 
-      expect { click_on "Pin!", match: :first }.to change { ProjectItem.count }.by(1)
+      expect { click_on "Send access request", match: :first }.to change { ProjectItem.count }.by(1)
     end
 
     scenario "I cannot order every 'open_access' type offer twice" do