+++ spec/policies/api/v1/message_policy_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Api::V1::MessagePolicy, type: :policy do
   let(:default_oms_admin) { create(:user) }
 
   let(:oms) { create(:oms, administrators: [oms_admin]) }
-  let(:default_oms) { create(:oms, default: true, administrators: [default_oms_admin]) }
+  let(:default_oms) { create(:default_oms, administrators: [default_oms_admin]) }
 
   let(:project_item1) { build(:project_item, offer: build(:offer, primary_oms: oms)) }
   let(:project_item2) { build(:project_item, offer: build(:offer, primary_oms: default_oms)) }